requirements.txt
Flask==3.0.2
boto3==1.34.69
python-dotenv==1.0.1
Werkzeug==3.0.1
gunicorn==21.2.0 